[#178] feat: 댓글 삭제시 확인창 표시

+++ frontend/src/components/comments/CommentDelete.jsx
@@ -3,10 +3,12 @@ import axios from "axios";
 function CommentDelete({ id, getComments }) {
 	const URL = process.env.REACT_APP_SERVER_URL;
 	const onClick = () => {
-		axios({
-			url: `${URL}/api/articles/comments/${id}`,
-			method: "delete",
-		}).then(() => getComments());
+		if (window.confirm("댓글을 삭제 하시겠습니까?")) {
+			axios({
+				url: `${URL}/api/articles/comments/${id}`,
+				method: "delete",
+			}).then(() => getComments());
+		}
 	};
 
 	return <button onClick={onClick}>삭제</button>;
